+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Experimental trade generate requirement hint
+
 ## [4.3.0] - 2024-08-18
 ### Added
 - Traditional Chinese translations